+++ frontend/src/components/reader/drawer/reader-settings-super-panel.vue
@@ -6,9 +6,9 @@
   <DownloadPanel />
 </template>
 <script>
-import DownloadPanel from "@/components/reader/download-panel.vue";
-import ReaderKeyboardShortcutsPanel from "@/components/reader/keyboard-shortcuts-panel.vue";
-import ReaderSettingsPanel from "@/components/reader/reader-settings-panel.vue";
+import DownloadPanel from "@/components/reader/drawer/download-panel.vue";
+import ReaderKeyboardShortcutsPanel from "@/components/reader/drawer/keyboard-shortcuts-panel.vue";
+import ReaderSettingsPanel from "@/components/reader/drawer/reader-settings-panel.vue";
 
 export default {
   name: "ReaderSettingsSuperPanel",