+++ frontend/src/components/reader/pages/pages-view.vue
@@ -1,20 +1,20 @@
 <template>
-  <PagesScrollerVertical v-if="vertical" :book="book" />
-  <PagesWindowHorizontal v-else :book="book" />
+  <PagesVerticalScroller v-if="vertical" :book="book" @click="$emit['click']" />
+  <PagesHorizontalWindow v-else :book="book" @click="$emit['click']" />
 </template>
 
 <script>
 import { mapActions, mapState } from "pinia";
 
-import PagesScrollerVertical from "@/components/reader/pages-scroller-vertical.vue";
-import PagesWindowHorizontal from "@/components/reader/pages-window-horizontal.vue";
+import PagesHorizontalWindow from "@/components/reader/pages/horizontal-window.vue";
+import PagesVerticalScroller from "@/components/reader/pages/vertical-scroller.vue";
 import { useReaderStore } from "@/stores/reader";
 
 export default {
   name: "PagesView",
   components: {
-    PagesWindowHorizontal,
-    PagesScrollerVertical,
+    PagesHorizontalWindow,
+    PagesVerticalScroller,
   },
   props: {
     book: { type: Object, required: true },
@@ -41,7 +41,7 @@ export default {
   created() {
     if (this.book.pk === this.storePk) {
       // Active Book
-      this.setActivePage(+this.$route.params.page);
+      this.setActivePage(+this.$route.params.page, true);
     }
   },
   methods: {